fix(projects and [slug]): code split for 3D model viewer, sizes prop for Images
